# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: ALAFIA Backend API
  version: 1.0.0
  description: |
    REST API for ALAFIA, a voice-first, safety-controlled healthcare navigation platform.
    AI structures user input; deterministic ALAFIA rules control triage severity.
servers:
  - url: http://localhost:5000/api/v1
    description: Local development
  - url: https://api.alafia.ai/api/v1
    description: Production placeholder
...
paths:
...
  /consultations:
    post:
      tags: [Consultation]
      summary: Create a text consultation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateConsultationRequest"
      responses:
        "201":
          description: Consultation created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConsultationCreatedResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
...
  /triage:
    post:
      tags: [Triage]
      summary: Run deterministic triage
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [consultationId]
              properties:
                consultationId: { type: string, format: uuid }
      responses:
        "200":
          description: Triage assessment completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TriageResponse"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /facilities:
    get:
      tags: [Facilities]
      summary: List all facilities
      responses:
        "200":
          description: Facilities retrieved
    post:
      tags: [Facilities]
      summary: Create a simulated facility
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FacilityInput"
      responses:
        "201":
          description: Facility created
        "400":
          $ref: "#/components/responses/ValidationError"
...
  /support-requests:
    post:
      tags: [Support Requests]
      summary: Create a trusted-contact support request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSupportRequest"
      responses:
        "201":
          description: Support request created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SupportRequestResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
...
  /webhooks/wema:
    post:
      tags: [Webhooks]
      summary: ALATPay webhook receiver (canonical)
      description: |
        Reserved for ALATPay callbacks. It is deliberately fail-closed until
        ALATPay publishes the exact webhook signature verification contract.
        Unauthenticated callbacks return 401 and never change payment state.
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, additionalProperties: true }
      responses:
        "401":
          description: Webhook authentication cannot be verified
        "503":
          description: Payment gateway is not configured
...
components:
...
  responses:
    ValidationError:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    Severity:
      type: string
      enum: [LOW, MEDIUM, HIGH, CRITICAL]
...
    CreateConsultationRequest:
      type: object
      required: [message]
      properties:
        sessionId: { type: string, format: uuid }
        inputType: { type: string, enum: [TEXT, VOICE], default: TEXT }
        message:
          {
            type: string,
            example: I have severe chest pain and difficulty breathing.,
          }
        language: { type: string, default: en-NG }
        location:
          type: object
          properties:
            latitude: { type: number }
            longitude: { type: number }
    FacilityInput:
      type: object
      required: [name, facilityType, latitude, longitude]
      properties:
        name: { type: string }
        facilityType:
          { type: string, enum: [hospital, clinic, urgent_care, pharmacy] }
        latitude: { type: number }
        longitude: { type: number }
        address: { type: string }
        phone: { type: string }
        email: { type: string, format: email }
        capabilities:
          type: array
          items: { type: string }
        emergencyCapable: { type: boolean, default: false }
        dataSource: { type: string, default: simulated }
        verificationStatus:
          { type: string, enum: [verified, unverified, pending] }
...
    ConsultationCreatedResponse:
      type: object
      properties:
        success: { type: boolean }
        data:
          type: object
          properties:
            consultation:
              type: object
              properties:
                id: { type: string, format: uuid }
                status: { type: string, example: ACTIVE }
                createdAt: { type: string, format: date-time }
            extraction:
              type: object
              properties:
                symptoms:
                  type: array
                  items:
                    type: object
                    properties:
                      name: { type: string }
                      confidence: { type: number, format: float }
            nextStep:
              type: object
              properties:
                type: { type: string }
                question: { type: string, nullable: true }
        message: { type: string }
    TriageResponse:
      type: object
      properties:
        success: { type: boolean }
        data:
          type: object
          properties:
            assessmentId: { type: string, format: uuid }
            severity: { $ref: "#/components/schemas/Severity" }
            confidence: { type: string, enum: [LOW, MEDIUM, HIGH] }
            redFlags:
              type: array
              items: { type: string }
            requiredCare:
              type: object
              properties:
                emergencyCare: { type: boolean }
                specialties:
                  type: array
                  items: { type: string }
                capabilities:
                  type: array
                  items: { type: string }
            guidance:
              type: object
              properties:
                title: { type: string }
                instructions:
                  type: array
                  items: { type: string }
            rulesVersion: { type: string }
        message: { type: string }
    ErrorResponse:
      type: object
      properties:
        success: { type: boolean, example: false }
        error:
          type: object
          properties:
            code: { type: string }
            message: { type: string }
            details:
              type: array
              items: {}
...
    CreateSupportRequest:
      type: object
      required: [patientRef, requestedAmountMinor]
      properties:
        patientRef: { type: string }
        consultationId: { type: string, format: uuid, nullable: true }
        requestedAmountMinor: { type: integer, format: int64, minimum: 1 }
        message: { type: string, nullable: true }
        expiresAt: { type: string, format: date-time, nullable: true }
        contacts:
          type: array
          maxItems: 20
          items:
            type: object
            properties:
              displayName: { type: string }
              contactChannel: { type: string, nullable: true }
    SupportRequestView:
      type: object
      properties:
        id: { type: string, format: uuid }
        shareToken: { type: string, format: uuid }
        status: { type: string, enum: [PENDING, PARTIALLY_FUNDED, FUNDED, EXPIRED, CANCELLED] }
        requestedAmountMinor: { type: integer, format: int64 }
        receivedAmountMinor: { type: integer, format: int64 }
        remainingAmountMinor: { type: integer, format: int64 }
        currency: { type: string, example: NGN }
        expiresAt: { type: string, format: date-time, nullable: true }
    SupportRequestResponse:
      type: object
      properties:
        success: { type: boolean, example: true }
        data: { $ref: "#/components/schemas/SupportRequestView" }
        message: { type: string }
...